# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckoutSessionResponse": {
        "properties": {
          "checkout_url": {
            "title": "Checkout Url",
            "type": "string"
          }
        },
        "required": [
          "checkout_url"
        ],
        "title": "CheckoutSessionResponse",
        "type": "object"
      },
      "ConsentGrantRequest": {
        "properties": {
          "consent_version": {
            "default": "1.0",
            "title": "Consent Version",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          }
        },
        "required": [
          "learner_id"
        ],
        "title": "ConsentGrantRequest",
        "type": "object"
      },
      "ConsentRevokeRequest": {
        "properties": {
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "reason": {
            "default": "guardian_request",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "learner_id"
        ],
        "title": "ConsentRevokeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAcceptedResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "operation"
        ],
        "title": "JobAcceptedResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered adaptive learning platform — Grade R to 7. CAPS-aligned. POPIA-compliant.",
    "title": "EduBoost SA V2",
    "version": "1.0.0-rc1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/admin/consent/trigger-renewal-reminders": {
      "post": {
        "description": "Queues the consent renewal reminder job as a FastAPI BackgroundTask.\nReturns 202 immediately; emails are dispatched asynchronously.\n\nAccess: Admin role required (RBAC enforced via ``require_role(\"Admin\")``\ndependency — wire once Task #15 RBAC is complete).",
        "operationId": "trigger_renewal_reminders_api_v2_admin_consent_trigger_renewal_reminders_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger POPIA consent renewal reminder emails (Admin only)",
        "tags": [
          "V2 Admin – Consent"
        ]
      }
    },
...
    "/api/v2/billing/checkout": {
      "post": {
        "operationId": "create_checkout_api_v2_billing_checkout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v2/billing/create-checkout-session": {
      "post": {
        "operationId": "create_checkout_api_v2_billing_create_checkout_session_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v2/billing/webhook": {
      "post": {
        "operationId": "stripe_webhook_api_v2_billing_webhook_post",
        "parameters": [
          {
            "in": "header",
            "name": "stripe-signature",
            "required": true,
            "schema": {
              "title": "Stripe-Signature",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stripe Webhook",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v2/consent/grant": {
      "post": {
        "operationId": "grant_consent_api_v2_consent_grant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Grant Consent Api V2 Consent Grant Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Grant Consent",
        "tags": [
          "POPIA Consent"
        ]
      }
    },
    "/api/v2/consent/revoke": {
      "post": {
        "operationId": "revoke_consent_api_v2_consent_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Revoke Consent Api V2 Consent Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Consent",
        "tags": [
          "POPIA Consent"
        ]
      }
    },
    "/api/v2/consent/status/{learner_id}": {
      "get": {
        "operationId": "consent_status_api_v2_consent_status__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Consent Status Api V2 Consent Status  Learner Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consent Status",
        "tags": [
          "POPIA Consent"
        ]
      }
    },
...
    "/v2/admin/consent/trigger-renewal-reminders": {
      "post": {
        "description": "Queues the consent renewal reminder job as a FastAPI BackgroundTask.\nReturns 202 immediately; emails are dispatched asynchronously.\n\nAccess: Admin role required (RBAC enforced via ``require_role(\"Admin\")``\ndependency — wire once Task #15 RBAC is complete).",
        "operationId": "trigger_renewal_reminders_v2_admin_consent_trigger_renewal_reminders_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger POPIA consent renewal reminder emails (Admin only)",
        "tags": [
          "V2 Admin – Consent"
        ]
      }
    },
...
    "/v2/billing/checkout": {
      "post": {
        "operationId": "create_checkout_v2_billing_checkout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/v2/billing/create-checkout-session": {
      "post": {
        "operationId": "create_checkout_v2_billing_create_checkout_session_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
    "/v2/billing/webhook": {
      "post": {
        "operationId": "stripe_webhook_v2_billing_webhook_post",
        "parameters": [
          {
            "in": "header",
            "name": "stripe-signature",
            "required": true,
            "schema": {
              "title": "Stripe-Signature",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stripe Webhook",
        "tags": [
          "billing"
        ]
      }
    },
    "/v2/consent/grant": {
      "post": {
        "operationId": "grant_consent_v2_consent_grant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Grant Consent V2 Consent Grant Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Grant Consent",
        "tags": [
          "POPIA Consent"
        ]
      }
    },
    "/v2/consent/revoke": {
      "post": {
        "operationId": "revoke_consent_v2_consent_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Revoke Consent V2 Consent Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Consent",
        "tags": [
          "POPIA Consent"
        ]
      }
    },
    "/v2/consent/status/{learner_id}": {
      "get": {
        "operationId": "consent_status_v2_consent_status__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Consent Status V2 Consent Status  Learner Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consent Status",
        "tags": [
          "POPIA Consent"
        ]
      }
    },
...
  },
...
}